minor fixes to l1t renaming
(cherry picked from commit 65f532cd8eaeb41191ca5c8df725cb85a05b3feb)
