+++ L1Trigger/L1TTrackMatch/python/L1TkMETAnalyser_cfi.py
@@ -1,10 +1,8 @@
 import FWCore.ParameterSet.Config as cms
-from L1Trigger.L1TTrackMatch.L1TrackerEtMissEmulatorProducer_cfi import l1tTrackerEmuEtMiss
-from L1Trigger.L1TTrackMatch.L1TrackerEtMissProducer_cfi import l1tTrackerEtMiss
 
 L1TkMETAnalyser = cms.EDAnalyzer('L1TkMETAnalyser',
-  TrackMETInputTag = cms.InputTag("l1tTrackerEtMiss",l1tTrackerEtMiss.L1MetCollectionName.value()),
-  TrackMETEmuInputTag = cms.InputTag("l1tTrackerEmuEtMiss",l1tTrackerEmuEtMiss.L1MetCollectionName.value()),
+  TrackMETInputTag = cms.InputTag("l1tTrackerEtMiss","l1tTrackerEtMiss"),
+  TrackMETEmuInputTag = cms.InputTag("l1tTrackerEmuEtMiss","L1TrackerEmuEtMiss"),
   TrackMETHWInputTag = cms.InputTag("GTTOutputFileReader"),
   HW_Analysis = cms.bool(False)
 )
